backend/services/transcriber/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-dotenv==1.0.0
aiofiles==23.2.1
numpy==1.26.3
httpx==0.27.0
pyjwt==2.9.0
cryptography==43.0.3
